+++ src/components/higlass/HiGlassSubscriber.js
@@ -1,7 +1,18 @@
 import React, {
   useMemo, useEffect, useCallback, Suspense,
 } from 'react';
+import register from 'higlass-register';
+import { ZarrMultivecDataFetcher } from 'higlass-zarr-datafetchers';
 import TitleInfo from '../TitleInfo';
+
+// Register the zarr-multivec plugin data fetcher.
+// References:
+// - https://github.com/higlass/higlass-register
+// - https://github.com/higlass/higlass-zarr-datafetchers
+register(
+  { dataFetcher: ZarrMultivecDataFetcher, config: ZarrMultivecDataFetcher.config },
+  { pluginType: 'dataFetcher' },
+);
 
 const HiGlassComponent = React.lazy(() => import('./HiGlass'));